Merged Location test with main branch

+++ LocationFinder/src/main/java/LocationFinder/services/LocationService.java
@@ -15,6 +15,7 @@ import com.sun.jdi.InvalidTypeException;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
+import java.util.*;
 
 @Service
 public class LocationService {
@@ -27,8 +28,8 @@ public class LocationService {
 
     public Location addLocation(Location loc) {
         loc.setClaim(false);
-        locRepository.save(loc);
-        return loc;
+        Location fullLocation = locRepository.save(loc);
+        return fullLocation;
     }
 
 
@@ -56,17 +57,21 @@ public class LocationService {
 
     public List<Location> getLocationByClaim(String claim_status) throws InvaildInputException {
         if(claim_status.toLowerCase().equals("claimed")) {
-            return locRepository.findByClaim(true);
+            List<Location> location_list = new LinkedList();
+            location_list = locRepository.findByClaim(true);
+            return location_list;
         }
         else if(claim_status.toLowerCase().equals("unclaimed")) {
-            return locRepository.findByClaim(false);
+            List<Location> location_list = new LinkedList();
+            location_list = locRepository.findByClaim(false);
+            return location_list;
         }
         else {
             throw new InvaildInputException("Please specify whether you are searching for claimed or unclaimed spots");
         }
     }
 
-    public Location updateLocClaim(Location loc, Boolean loc_claim) {
+    public Location updateLocClaim(Location loc, Boolean loc_claim) throws IllegalArgumentException{
         Location updatedLoc = new Location(loc.getName(), loc.getArea(), loc.getCost());
         updatedLoc.setId(loc.getId());
         updatedLoc.setClaim(loc_claim);
@@ -82,8 +87,16 @@ public class LocationService {
         return updatedLoc;
     }
 
-    public void deleteLocationById(Integer id) {
+    public void deleteLocationById(Integer id) throws NotFoundException{
+        if(locRepository.existsById(id)) {
+
         locRepository.deleteById(id);
+        }
+        else {
+            throw new NotFoundException("There is no location with that id");
+
+        }
+
     }
 }
